Add Linux legacy routing if ip rule functionality is not available
